requirements.txt
google-adk
pdfplumber
chromadb
tiktoken
gradio
streamlit
pydantic
python-dotenv
google-generativeai
